[hotfix] Move infra from flink-ml-lib to flink-ml-core
