+++ flink-ml-core/src/main/java/org/apache/flink/ml/common/datastream/MapPartitionFunctionWrapper.java
