+++ flink-ml-lib/pom.xml
@@ -99,13 +99,6 @@ under the License.
       <version>${flink.version}</version>
       <scope>test</scope>
     </dependency>
-    <dependency>
-      <groupId>org.mockito</groupId>
-      <artifactId>mockito-core</artifactId>
-      <version>2.21.0</version>
-      <type>jar</type>
-      <scope>test</scope>
-    </dependency>
   </dependencies>
 
 </project>
